requirements.txt
streamlit==1.48.1
openai
langchain
langchain-openai
python-dotenv
numpy

